move contact_system into events module
